pyproject.toml
[tool.poetry]
name = "assignment"
version = "0.1.0"
description = ""
authors = ["Neeraj319 <itskop520@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
python-dotenv = "^0.20.0"
PyYAML = "^6.0"
uritemplate = "^4.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
